our-journey/requirements.txt
aws-cdk-lib>=2.211.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
cdklabs.generative-ai-cdk-constructs==0.1.312
